Reduced memory usage of canvas search

+++ code/new_code/Canvas.hh
@@ -12,7 +12,7 @@
 
 
 struct CanvasCode {
-    std::vector<int> code;
+    std::vector<int8_t> code;
 
     CanvasCode();
